requirements.txt
flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
PyMySQL==1.1.1
WTForms==3.1.2
Flask-SQLAlchemy==3.1.1
openpyxl==3.1.5
